+++ src/main/java/rbasamoyai/ogden/OgdenMod.java
@@ -8,10 +8,12 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.level.LevelAccessor;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.event.OnDatapackSyncEvent;
+import net.minecraftforge.event.world.WorldEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.common.Mod;
@@ -19,6 +21,7 @@ import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import rbasamoyai.ogden.ammunition.AmmunitionPropertiesHandler;
+import rbasamoyai.ogden.firearms.config.FirearmAcceptedAmmunitionHandler;
 import rbasamoyai.ogden.index.OgdenAmmoGroup;
 import rbasamoyai.ogden.index.OgdenBaseGroup;
 import rbasamoyai.ogden.index.OgdenEntityTypes;
@@ -48,6 +51,7 @@ public class OgdenMod {
 
         forgeBus.addListener(this::registerResourceListeners);
         forgeBus.addListener(this::onDatapackSync);
+        forgeBus.addListener(this::onLoadLevel);
 
         OgdenConfigs.registerConfigs();
 
@@ -66,15 +70,27 @@ public class OgdenMod {
 
     private void registerResourceListeners(final AddReloadListenerEvent evt) {
         evt.addListener(AmmunitionPropertiesHandler.ReloadListener.INSTANCE);
+        evt.addListener(FirearmAcceptedAmmunitionHandler.ReloadListener.INSTANCE);
     }
 
     public void onDatapackSync(final OnDatapackSyncEvent evt) {
+        FirearmAcceptedAmmunitionHandler.loadTags();
+
         if (evt.getPlayer() == null) {
             MinecraftServer server = evt.getPlayerList().getServer();
-            AmmunitionPropertiesHandler.syncToServer(server);
+            AmmunitionPropertiesHandler.syncToAll(server);
+            FirearmAcceptedAmmunitionHandler.syncToAll(server);
         } else {
             ServerPlayer player = evt.getPlayer();
             AmmunitionPropertiesHandler.syncToPlayer(player);
+            FirearmAcceptedAmmunitionHandler.syncToPlayer(player);
+        }
+    }
+
+    public void onLoadLevel(final WorldEvent.Load evt) {
+        LevelAccessor level = evt.getWorld();
+        if (level.getServer() != null && !level.isClientSide() && level.getServer().overworld() == level) {
+            FirearmAcceptedAmmunitionHandler.loadTags();
         }
     }